+++ management/views/preparation.py
@@ -1,3 +1,4 @@
+from django.core.exceptions import PermissionDenied
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from management.utils.pagination import Pagination
@@ -8,6 +9,20 @@ from management import models
 
 def preparation(request, _type):
     """已有制剂的已完成，待开发，开发中进度表"""
+
+    # 权限映射
+    permission_map = {
+        'completed': 'management.view_existingproductcompleted',
+        'ing': 'management.view_existingformulationdeveloping',
+        'todev': 'management.view_existingformulationtodevelop',
+    }
+
+    # 检查权限
+    user_has_permission = request.user.has_perm(permission_map.get(_type, ''))
+    if not user_has_permission:
+        raise PermissionDenied
+
+    # 模型映射
     model_map = {
         'completed': models.ExistingProductCompleted,
         'ing': models.ExistingFormulationDeveloping,
@@ -44,12 +59,35 @@ def preparation(request, _type):
         'page_string': page_object.page_string,
         'value': value,
         'type': _type,
+        # 为每种类型和操作组合检查权限
+        'can_add_completed': request.user.has_perm('management.add_existingproductcompleted'),
+        'can_edit_completed': request.user.has_perm('management.change_existingproductcompleted'),
+        'can_delete_completed': request.user.has_perm('management.delete_existingproductcompleted'),
+
+        'can_add_ing': request.user.has_perm('management.add_existingformulationdeveloping'),
+        'can_edit_ing': request.user.has_perm('management.change_existingformulationdeveloping'),
+        'can_delete_ing': request.user.has_perm('management.delete_existingformulationdeveloping'),
+
+        'can_add_todev': request.user.has_perm('management.add_existingformulationtodevelop'),
+        'can_edit_todev': request.user.has_perm('management.change_existingformulationtodevelop'),
+        'can_delete_todev': request.user.has_perm('management.delete_existingformulationtodevelop'),
     }
     return render(request, 'preparation.html', context)
 
 
 def preparation_add(request, _type):
     """已有制剂（待开发，开发中，已完成）添加"""
+    # 权限映射
+    permission_map = {
+        'completed': 'management.add_existingproductcompleted',
+        'ing': 'management.add_existingformulationdeveloping',
+        'todev': 'management.add_existingformulationtodevelop',
+    }
+    # 检查权限
+    user_has_permission = request.user.has_perm(permission_map.get(_type, ''))
+    if not user_has_permission:
+        raise PermissionDenied
+
     form_cls = {
         'completed': ExistingProductCompleted_form,
         'ing': ExistingFormulationDeveloping_form,
@@ -68,6 +106,17 @@ def preparation_add(request, _type):
 
 def preparation_edit(request, _type, _id):
     """已有制剂开发表"""
+    # 权限映射
+    permission_map = {
+        'completed': 'management.change_existingproductcompleted',
+        'ing': 'management.change_existingformulationdeveloping',
+        'todev': 'management.change_existingformulationtodevelop',
+    }
+    # 检查权限
+    user_has_permission = request.user.has_perm(permission_map.get(_type, ''))
+    if not user_has_permission:
+        raise PermissionDenied
+
     form_cls = {
         'completed': ExistingProductCompleted_form,
         'ing': ExistingFormulationDeveloping_form,
@@ -91,6 +140,17 @@ def preparation_edit(request, _type, _id):
 
 
 def preparation_delete(request, _type, _id):
+    # 权限映射
+    permission_map = {
+        'completed': 'management.delete_existingproductcompleted',
+        'ing': 'management.delete_existingformulationdeveloping',
+        'todev': 'management.delete_existingformulationtodevelop',
+    }
+    # 检查权限
+    user_has_permission = request.user.has_perm(permission_map.get(_type, ''))
+    if not user_has_permission:
+        raise PermissionDenied
+
     """删除已有制剂开发表"""
     if _type == 'completed':
         models.ExistingProductCompleted.objects.filter(serial_number=_id).delete()
